Major incompatible renaming: Copy -> Flat

+++ mem_dbg/tests/test_mem_size.rs
@@ -16,7 +16,7 @@ use std::sync::atomic::AtomicU64;
 #[allow(dead_code)]
 #[derive(MemSize)]
 #[cfg_attr(feature = "std", derive(MemDbg))]
-#[move_type]
+#[mem_size_rec]
 enum TestEnum {
     Unit,
     Unit2(),
@@ -44,7 +44,7 @@ fn test_vec_capacity() {
 fn test_vec_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[move_type]
+    #[mem_size_rec]
     struct NewType(usize);
 
     assert_eq!(
@@ -57,7 +57,7 @@ fn test_vec_copy_or_not() {
 fn test_boxed_slice_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[move_type]
+    #[mem_size_rec]
     struct NewType(usize);
 
     assert_eq!(
@@ -74,7 +74,7 @@ fn test_boxed_slice_copy_or_not() {
 fn test_slice_copy_or_not() {
     #[derive(Clone, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[move_type]
+    #[mem_size_rec]
     struct NewType(usize);
 
     assert_eq!(
@@ -93,7 +93,7 @@ fn test_slice_copy_or_not() {
 fn test_array_copy_or_not() {
     #[derive(Clone, Copy, MemSize)]
     #[cfg_attr(feature = "std", derive(MemDbg))]
-    #[copy_type]
+    #[mem_size_flat]
     struct NewType(usize);
 
     assert_eq!(
@@ -107,7 +107,7 @@ fn test_array_copy_or_not() {
 #[test]
 fn test_empty_struct() {
     #[derive(MemSize, Clone, Copy)]
-    #[copy_type]
+    #[mem_size_flat]
     struct Data {}
     let v = Data {};
     assert_eq!(v.mem_size(SizeFlags::default()), 0);
@@ -140,7 +140,7 @@ fn test_struct() {
 #[test]
 fn test_empty_tuple_struct() {
     #[derive(MemSize)]
-    #[move_type]
+    #[mem_size_rec]
     struct Data();
     let v = Data();
     assert_eq!(v.mem_size(SizeFlags::default()), 0);
@@ -168,13 +168,13 @@ fn test_tuple_struct() {
 fn test_padding() {
     assert_eq!((0_u8, 0_u64).mem_size(SizeFlags::default()), 16);
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[move_type]
+    #[mem_size_rec]
     struct TuplePadded((u8, u64));
     let v = TuplePadded((0, 0));
     assert_eq!(v.mem_size(SizeFlags::default()), 16);
 
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[move_type]
+    #[mem_size_rec]
     struct StructPadded(u8, u64);
     let v = StructPadded(0, 0);
     assert_eq!(v.mem_size(SizeFlags::default()), 16);
@@ -269,7 +269,7 @@ fn test_exotic() {
     );
 
     #[cfg_attr(feature = "derive", derive(MemSize))]
-    #[move_type]
+    #[mem_size_rec]
     enum Data1 {
         A,
         B,
@@ -325,7 +325,7 @@ fn test_phantom() {
     struct Dummy();
     #[cfg_attr(feature = "derive", derive(MemSize))]
     #[cfg_attr(all(feature = "std", feature = "derive"), derive(MemDbg))]
-    #[move_type]
+    #[mem_size_rec]
     struct Example<A>(PhantomData<A>);
 
     Example::<Dummy>(PhantomData)
@@ -365,7 +365,7 @@ fn test_array_u8() {
 #[cfg(feature = "std")]
 fn test_array_empty_struct() {
     #[derive(MemSize, MemDbg, Clone, Copy)]
-    #[copy_type]
+    #[mem_size_flat]
     struct Dummy;
     let data = [Dummy; 10];
     assert_eq!(data.mem_size(SizeFlags::default()), 0);
@@ -395,7 +395,7 @@ fn test_slice_u8() {
 #[test]
 fn test_slice_empty_struct() {
     #[derive(MemSize, MemDbg, Clone, Copy)]
-    #[copy_type]
+    #[mem_size_flat]
     struct Dummy;
     let data = [Dummy; 10].as_slice();
     assert_eq!((*data).mem_size(SizeFlags::default()), 0);
@@ -545,18 +545,18 @@ fn test_vec_slice_i32() {
         "Expected mutable slice deep size to be identical to non mutable slice deep size"
     );
 
-    assert_eq!(non_mutable_slice_shallow_size, size_of::<&[i64]>());
+    assert_eq!(non_mutable_slice_shallow_size, size_of::<&[i32]>());
 
-    assert_eq!(mutable_slice_shallow_size, size_of::<&mut [i64]>());
+    assert_eq!(mutable_slice_shallow_size, size_of::<&mut [i32]>());
 
     assert_eq!(
         non_mutable_slice_deep_size,
-        size_of::<&[i64]>() + size_of_non_mutable_slice
+        size_of::<&[i32]>() + size_of_non_mutable_slice
     );
 
     assert_eq!(
         mutable_slice_deep_size,
-        size_of::<&mut [i64]>() + size_of_mutable_slice
+        size_of::<&mut [i32]>() + size_of_mutable_slice
     );
 }
 
@@ -665,12 +665,12 @@ fn test_array_slice_i32() {
 
     assert_eq!(
         non_mutable_slice_deep_size,
-        core::mem::size_of::<&[i64]>() + size_of_non_mutable_slice
+        core::mem::size_of::<&[i32]>() + size_of_non_mutable_slice
     );
 
     assert_eq!(
         mutable_slice_deep_size,
-        core::mem::size_of::<&mut [i64]>() + size_of_mutable_slice
+        core::mem::size_of::<&mut [i32]>() + size_of_mutable_slice
     );
 }
